report/python/pilot/requirements.txt
pyyaml
prometheus-api-client
pandas
numpy
matplotlib
jinja2
pyppeteer
Pillow